chore(ruff): bring codebase up to ruff 0.15.12 lint standards
Ruff has shipped multiple new lint rules since the 0.6 baseline pinned in
pyproject.toml. Running the latest version flagged 94 issues across the
codebase that previously slipped past — blocking dependabot dep-bump PRs
(#22 ruff, #23 bandit, #24 setuptools, #25 pyarrow) which all pull a
fresh ruff into the CI environment.

Two-pronged cleanup:

**Auto-fix (65 issues, all behind --unsafe-fixes):**
- B905  zip-without-explicit-strict (×21)
- RUF005 collection literal concatenation (×7)
- RUF046 unnecessary cast-to-int (×6)
- B007 unused loop variable (×4)
- F841 unused variable (×4)
- C408 unnecessary collection call, RUF007 zip→pairwise, SIM103 needless
  bool, RUF015 unnecessary iterable allocation, etc.

**Manual / config (29 remaining):**
- PLR0911 too-many-returns: per-function noqa with comment for the 4 active
  files (edgar_companyfacts._resolve_components, simfin._ps_ratio_pit,
  fetch_survivorship_ohlcv.should_fetch, theme_classifier.classify_theme).
  Each is a sequential dispatch / partial-failure resolver where early
  returns are clearer than nested-if. Closed-layer files inherit a
  package-level ignore via the new alphalens/archive/** rule.
- B904 raise-from in two places (archive/guru/llm_scorer + cli/queue) —
  added `from None` to express intentional discard of upstream context.
- E741 ambiguous `l` in OHLCV unpacking (lean_csv_writer + test) —
  per-line noqa with comment on the convention.
- New ignores in pyproject:
  * Global PLC2401 (non-ASCII names): math notation policy per CLAUDE.md
    (α_q, ρ_t, β_smb in research scripts)
  * `tests/*`: N802 (SEC Form 4 codes preserve case), RUF012 (mutable
    fixtures), B035 (static-key fixtures), SIM115/SIM117 (tempfile
    pattern), PLW1510 (subprocess returncode is the test's assertion)
  * `alphalens/archive/**`: PLW0603, PLR0911 — closed code per ADR 0005,
    not refactoring closed modules
- Updated `alphalens/watchdog_lock.py` per-file-ignore path to
  `alphalens/watchdog/lock.py` after the Phase 5 reorg move.

Pre-existing issues caught by the new rules; no logic changes.

`ruff check .` reports `All checks passed!` after the changes. 1425 tests
pass; ruff format reports 347/347 files already formatted.

Unblocks dependabot PRs #22-25.

+++ scripts/experiment_vol_target_overlay.py
@@ -93,7 +93,7 @@ def assess_overlay(
     if raw_rets.empty:
         return {"n": 0}
 
-    rebalances_per_year = max(1, int(round(252 / max(1, rebalance_stride))))
+    rebalances_per_year = max(1, round(252 / max(1, rebalance_stride)))
 
     snapshots = [list(r.top_n_tickers) for r in report.rebalance_results]
     overlay_stats = compute_overlay_stats(
@@ -213,7 +213,7 @@ def main() -> int:
     histories = load_cached_histories([*full_universe, args.benchmark], _PRICES_DIR)
     history_store = HistoryStore(histories)
 
-    rebalances_per_year = max(1, int(round(252 / args.rebalance_stride)))
+    rebalances_per_year = max(1, round(252 / args.rebalance_stride))
 
     periods = [
         (f"IS {args.is_start.year}-{args.is_end.year}", args.is_start, args.is_end),
